op snapshot sync log print hex string instead of number array
