+++ crates/cfx_types/Cargo.toml
@@ -10,6 +10,7 @@ serde_derive = "1.0"
 serde = "1.0"
 rlp = "0.4.0"
 rlp_derive = { git = "https://github.com/Conflux-Chain/conflux-parity-deps.git", rev = "1597a9cab02343eb2322ca0ac58d39b64e3f42d1"  }
+hex = "0.4.3"
 
 [features]
 storage_benchmark_no_account_space_check = []
